Add localised queue signalling on free slots
Instead of a global free slot condition variable, be smarter and signal only the connected start processor of the free slot.

This on its own doesn't do very much and saves a very small number of CPU cycles, but it does enable the flexibility to pass on this signal to the internal job queue condition variable. This is useful as since we added queue categorisation, a waiting queue no longer means it is empty and it could just mean there were no free slots for that specific subqueue, so we should signal it when there's a potential change in job slot distribution.